requirements.txt
flask
flask_cors
gunicorn
psycopg2-binary